refactor: collapse loreStore from 19 files to 8
The lore store was split into 19 files following a layer-cake pattern:
- a public root file
- an aggregator (actions.facts.ts)
- per-concern action wrappers (actions.facts.maintenance.ts, actions.aliases.ts,
  actions.import-export.ts, etc) that each built a "context" object
- the actual impl files (loreStore.utils.ts, loreStore.aliases.ts, etc) that
  consumed those contexts
plus tiny state/type-alias files.

Every action wrapper did the same job: import the impl, construct a context
of store getters, call the impl. With set/get already in scope inside the
store factory, the indirection paid for nothing.

Consolidated into one domain file per concern:
  loreStore.ts             store + state + persist (was: ts + state.ts + actions.types.ts)
  loreStore.facts.ts       base + facts CRUD + validation + maintenance + utils
  loreStore.deduplication.ts  dedup + entity resolution + reference updates
  loreStore.aliases.ts     alias actions + impl
  loreStore.import-export.ts  import/export actions + impl
  loreStore.usage.ts       usage tracking (renamed from actions.usage.ts)
  loreStore.extraction.ts  structured lore extraction (unchanged)
  loreStore.helpers.ts     cross-cutting helpers (generateLoreKey etc, unchanged)

Public API (useLoreStore) is unchanged - only internal file organisation
changed. The Impl-with-context exports in loreStore.deduplication.ts are
preserved so the existing deduplication unit tests keep working.

19 files -> 8 files, ~360 net LOC removed (mostly wrapper boilerplate).
All 182 loreStore tests and 342 narrative tests pass.
