+++ src/state/loreStore.ts
@@ -12,18 +12,34 @@ import type {
   LoreUsageEvent,
   LoreUsageSource,
   LoreUsageStats,
+  DuplicateMatch,
 } from '../types/lore.types';
 import type { EntityID } from '../types/common.types';
 import { createIndexedDBStorage } from './persistence';
 import { UserFriendlyError } from '@/lib/utils/errorUtils';
 import { CrudStore } from './createCrudStore';
-import type { DuplicateMatch } from '../types/lore.types';
-import { getInitialState, type FactHistory } from './loreStore.state';
-import { createLoreBaseActions } from './loreStore.actions.base';
-import { createLoreFactActions } from './loreStore.actions.facts';
-import { createLoreAliasActions } from './loreStore.actions.aliases';
-import { createLoreUsageActions } from './loreStore.actions.usage';
-import { createLoreImportExportActions } from './loreStore.actions.import-export';
+import { createLoreFactActions } from './loreStore.facts';
+import { createLoreAliasActions } from './loreStore.aliases';
+import { createLoreImportExportActions } from './loreStore.import-export';
+import { createLoreUsageActions } from './loreStore.usage';
+import { createLoreDeduplicationActions } from './loreStore.deduplication';
+
+export interface FactHistory {
+  factId: EntityID;
+  versions: LoreFact[];
+}
+
+export const getInitialState = () => ({
+  facts: {} as Record<EntityID, LoreFact>,
+  entities: {} as Record<EntityID, LoreFact>,
+  factHistory: {} as Record<EntityID, FactHistory>,
+  mergeAuditLog: [] as LoreMergeAuditEntry[],
+  loreUsage: {} as Record<EntityID, LoreUsageStats>,
+  loreUsageEvents: [] as LoreUsageEvent[],
+  currentEntityId: null as EntityID | null,
+  error: null as UserFriendlyError | null,
+  loading: false,
+});
 
 export interface LoreStore extends CrudStore<LoreFact> {
   facts: Record<EntityID, LoreFact>;
@@ -57,7 +73,9 @@ export interface LoreStore extends CrudStore<LoreFact> {
   exportFacts: (worldId: EntityID) => string;
   importFacts: (worldId: EntityID, jsonData: string) => void;
   getFactHistory: (id: EntityID) => LoreFact[];
-  validateFact: (fact: Partial<{ key: string; value: string; category: LoreCategory; worldId: EntityID }>) => boolean;
+  validateFact: (
+    fact: Partial<{ key: string; value: string; category: LoreCategory; worldId: EntityID }>
+  ) => boolean;
   validateKey: (key: string) => boolean;
   getLoreContext: (worldId: EntityID, sessionId?: EntityID, limit?: number) => LoreContext;
   addStructuredLore: (extraction: StructuredLoreExtraction, worldId: EntityID, sessionId?: EntityID) => void;
@@ -67,8 +85,15 @@ export interface LoreStore extends CrudStore<LoreFact> {
   findEntityByAnyName: (name: string, worldId: EntityID) => LoreFact | null;
   scanForDuplicates: (worldId: EntityID, category?: LoreCategory) => Promise<DuplicateMatch[]>;
   mergeFacts: (primaryId: EntityID, secondaryId: EntityID) => void;
-  checkDuplicateBeforeCreate: (value: string, category: LoreCategory, worldId: EntityID) => Promise<DuplicateMatch[]>;
-  findPotentialEntityMatches: (worldId: EntityID, options?: { minConfidence?: number; category?: LoreCategory }) => EntityMatch[];
+  checkDuplicateBeforeCreate: (
+    value: string,
+    category: LoreCategory,
+    worldId: EntityID
+  ) => Promise<DuplicateMatch[]>;
+  findPotentialEntityMatches: (
+    worldId: EntityID,
+    options?: { minConfidence?: number; category?: LoreCategory }
+  ) => EntityMatch[];
   getMergeAuditLog: () => LoreMergeAuditEntry[];
   recordLoreUsage: (input: {
     worldId: EntityID;
@@ -84,18 +109,20 @@ export interface LoreStore extends CrudStore<LoreFact> {
     source?: LoreUsageSource;
   }) => void;
   clearLoreUsage: (worldId?: EntityID) => void;
+  reset: () => void;
 }
 
 export const useLoreStore = create<LoreStore>()(
   persist(
     (set, get) => ({
       ...getInitialState(),
-      ...createLoreBaseActions(set, get),
       ...createLoreFactActions(set, get),
       ...createLoreAliasActions(set, get),
       ...createLoreImportExportActions(set, get),
       ...createLoreUsageActions(set, get),
+      ...createLoreDeduplicationActions(set, get),
       getMergeAuditLog: () => get().mergeAuditLog,
+      reset: () => set(getInitialState()),
     }),
     {
       name: 'lore-store',
